resolved PR feedback
